Refactor concolic variables and handle PutField
Concolic variables are now organized differently. Each input is an
instance of Concolic, which has two fields: concVal and symVal, which
contain the concrete and symbolic instances of the variable,
respectively.

This allows us to separate the concrete information from the symbolic
information more cleanly. The classes in symbolic.js no longer contain
or interact with concrete information, so we can manipulate them as if
they were abstract formulae. This also means that concrete operations
are now performed eagerly, and only once, rather than on every call to
SymbolicValue.eval().

This also allows us to modify the underlying symbolic value, which has
been used to perform a rudimentary implementation of PutField by
chaining them onto the symbolic objects whenever a property set occurs.